# protocol/patricia: use pointer instead of raw hash