Passing the raw hash allocates a new hash on every call to insert, which
is itself a recursive call. In a recent run of benchcore, insert
accounted for ~3% of all allocations, about half of which were from
passing the raw hash.

```
benchmark              old ns/op     new ns/op     delta
BenchmarkInserts-8     29747152      26422206      -11.18%

benchmark              old allocs     new allocs     delta
BenchmarkInserts-8     283503         161759         -42.94%

benchmark              old bytes     new bytes     delta
BenchmarkInserts-8     15546252      11647064      -25.08%
```