@erykwalder 2016-11-29
A gist of the mem profile from the function before:
https://gist.github.com/erykwalder/4c6001820988c256781392e77a9789a1